chore: rename directory names as actual name for Input/Table
