+++ client/components/page/Admin/Invoice/InvoiceList.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useColorModeValue, Text, Flex } from '@chakra-ui/react';
-import { SimpleTable } from '@/components/shared/Tabletmp';
+import { SimpleTable } from '@/components/shared/Table';
 import { InvoiceListNextPageProps } from 'pages/admin/invoice';
 import { currencyFormat } from '@/components/function/utils/format';
